+++ app/uses/layout.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
-import MainLayout from '../../components/layouts/MainLayout';
+import MainLayout from '../components/main-layout';
 
 export default function Layout({ children }: { children: ReactNode }) {
   return <MainLayout>{children}</MainLayout>;
