--- components/layouts/MDX/ListLayout.tsx
@@ -1,62 +1,0 @@
-'use client';
-
-import Pagination from '@/components/Pagination';
-import PostCard from '@/components/PostCard';
-import { ComponentProps, useState } from 'react';
-import { BlogPost } from '../../../app/blog/utils';
-
-interface Props {
-  posts: BlogPost[];
-  title: string;
-  initialDisplayPosts?: BlogPost[];
-  pagination?: ComponentProps<typeof Pagination>;
-}
-
-export default function ListLayout({ posts, title, initialDisplayPosts = [], pagination }: Props) {
-  const [searchValue, setSearchValue] = useState('');
-  const filteredBlogPosts = posts.filter((post) => {
-    const searchContent = post.metadata.title + post.metadata.summary;
-    return searchContent.toLowerCase().includes(searchValue.toLowerCase());
-  });
-
-  // If initialDisplayPosts exist, display it if no searchValue is specified.
-  const displayPosts =
-    initialDisplayPosts.length > 0 && !searchValue ? initialDisplayPosts : filteredBlogPosts;
-
-  return (
-    <>
-      <div className="space-y-2 rounded-lg pt-8 pb-3 md:space-y-5">
-        <h1 className="text-3xl font-extrabold leading-9 tracking-tight text-gray-900 dark:text-gray-100 sm:text-4xl sm:leading-10 md:text-5xl md:leading-14">
-          {title}
-        </h1>
-        <div className="relative max-w-full">
-          <input
-            aria-label="Search articles"
-            type="text"
-            onChange={(e) => setSearchValue(e.target.value)}
-            placeholder="Search articles"
-            className="block w-full rounded-md border-0 bg-gray-200 bg-opacity-50 px-4 py-3 text-gray-900 focus:border-sky-500 focus:ring-sky-500 dark:border-gray-900 dark:bg-gray-800 dark:text-gray-100"
-          />
-          <svg
-            className="absolute right-3 top-3 h-6 w-6 text-gray-400 dark:text-gray-300"
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 24 24"
-            stroke="currentColor"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth={2}
-              d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-            />
-          </svg>
-        </div>
-        <PostCard posts={displayPosts} />
-      </div>
-      {pagination && pagination.totalPages > 1 && !searchValue && (
-        <Pagination currentPage={pagination.currentPage} totalPages={pagination.totalPages} />
-      )}
-    </>
-  );
-}
